# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AirFlex API
  description: |
    REST API for the AirFlex P2P airtime and data marketplace built on Stellar.

    All endpoints (except `/health`) are prefixed with `/api/v1/`.
    Every response includes an `X-Api-Version: 1` header.

    See [CONTRIBUTING.md](../CONTRIBUTING.md#api-versioning-policy) for the full
    versioning policy.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Local development server
security: []
...
paths:
...
  /api/v1/auth/request-otp:
    post:
      tags: [Auth]
      summary: Request an OTP sent via SMS
      operationId: requestOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestOtpInput"
      responses:
        "200":
          description: OTP sent successfully
          headers:
            X-Api-Version:
              $ref: "#/components/headers/X-Api-Version"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: OTP sent
        "400":
          $ref: "#/components/responses/ValidationError"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify OTP and receive JWT
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: OTP verified — JWT returned
          headers:
            X-Api-Version:
              $ref: "#/components/headers/X-Api-Version"
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: Signed JWT to use in subsequent requests
                  user:
                    $ref: "#/components/schemas/UserProfile"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/trades:
    get:
      tags: [Trades]
      summary: List paginated active trade offers
      operationId: listTrades
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated list of active trades
          headers:
            X-Api-Version:
              $ref: "#/components/headers/X-Api-Version"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTradesResponse"
        "400":
          $ref: "#/components/responses/ValidationError"

    post:
      tags: [Trades]
      summary: Create a new trade offer
      operationId: createTrade
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTradeInput"
      responses:
        "201":
          description: Trade offer created
          headers:
            X-Api-Version:
              $ref: "#/components/headers/X-Api-Version"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/profile:
    get:
      tags: [Profile]
      summary: Get authenticated user's profile
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          headers:
            X-Api-Version:
              $ref: "#/components/headers/X-Api-Version"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/profile/trades:
    get:
      tags: [Profile]
      summary: Get authenticated user's trade history
      operationId: getProfileTrades
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, buying, selling]
            default: all
      responses:
        "200":
          description: Paginated list of user's trades
          headers:
            X-Api-Version:
              $ref: "#/components/headers/X-Api-Version"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTradesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    X-Api-Version:
      description: Current API version number
      schema:
        type: string
        example: "1"
  schemas:
...
    RequestOtpInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          description: |
            Nigerian local format (0XXXXXXXXXX) or international E.164 (+2348XXXXXXXXX).
          example: "+2348012345678"
    VerifyOtpInput:
      type: object
      required: [phone, otp]
      properties:
        phone:
          type: string
          example: "+2348012345678"
        otp:
          type: string
          minLength: 6
          maxLength: 6
          pattern: '^\d{6}$'
          example: "123456"
    CreateTradeInput:
      type: object
      required: [assetType, amount, expiresInHours]
      properties:
        assetType:
          type: string
          pattern: '^[A-Za-z0-9_-]{1,50}$'
          example: MTN_AIRTIME
        amount:
          type: number
          minimum: 0.000001
          example: 500
        expiresInHours:
          type: integer
          minimum: 1
          maximum: 168
          example: 24
...
    Trade:
      type: object
      properties:
        id:
          type: string
          format: uuid
        seller_id:
          type: string
          format: uuid
        buyer_id:
          type: string
          format: uuid
          nullable: true
        asset_type:
          type: string
          example: MTN_AIRTIME
        amount:
          type: number
          example: 500
        status:
          type: string
          enum: [Active, Locked, Completed, Disputed, Expired]
        contract_listing_id:
          type: string
          nullable: true
        escrow_tx_hash:
          type: string
          nullable: true
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TradeResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Trade"
    PaginatedTradesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
        stellarPublicKey:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  responses:
    ValidationError:
      description: Request body or query parameters failed validation
      headers:
        X-Api-Version:
          $ref: "#/components/headers/X-Api-Version"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    UnprocessableEntity:
      description: Unprocessable entity
      headers:
        X-Api-Version:
          $ref: "#/components/headers/X-Api-Version"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid JWT
      headers:
        X-Api-Version:
          $ref: "#/components/headers/X-Api-Version"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Missing or invalid Authorization header
...